+++ examples/example-app-v12-monorepo/jest-isolated.config.js
@@ -1,4 +1,0 @@
-/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
-  projects: ['<rootDir>/projects/app1/jest-isolated.config.js', '<rootDir>/projects/app2/jest-isolated.config.js'],
-};