+++ src/compiler/ng-jest-compiler.ts
@@ -1,6 +1,6 @@
-import { TsCompiler } from 'ts-jest/dist/compiler/ts-compiler';
-import { ConfigSet } from 'ts-jest/dist/config/config-set';
-import type { TsJestAstTransformer } from 'ts-jest/dist/types';
+import type { TsJestAstTransformer } from 'ts-jest';
+import { TsCompiler } from 'ts-jest/dist/legacy/compiler/ts-compiler';
+import { ConfigSet } from 'ts-jest/dist/legacy/config/config-set';
 import type * as ts from 'typescript';
 
 import { constructorParametersDownlevelTransform } from '../transformers/downlevel-ctor';